Add aggregate stats logback logging to server.
This is so that during deployment to a server such as Heroku,
we can separate out these meaingful stats about whether the server
is alive and being sued from other stdout logs from Gdx.app.log/debug.
